% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmFromDuckDB.R
\name{cdmFromDuckDB}
\alias{cdmFromDuckDB}
\title{Title}
\usage{
cdmFromDuckDB(
  con,
  cdmSchema = "main",
  cdmPrefix = "",
  writeSchema = "main",
  writePrefix = "",
  achillesSchema = NULL,
  achillesPrefix = "",
  cohortTables = character(),
  otherTables = character(),
  cdmVersion = NULL,
  cdmName = NULL,
  .softValidation = FALSE
)
}
\arguments{
\item{con}{A DuckDB connection or path to a duckdb file.}

\item{.softValidation}{}
}
\description{
Title
}
