% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdbSource.R
\name{duckdbSource}
\alias{duckdbSource}
\title{Create a `DuckDB` source}
\usage{
duckdbSource(con = NULL, writeSchema = NULL, writePrefix = NULL)
}
\arguments{
\item{con}{A DuckDB connection or path to a duckdb file. If NULL a temp connection will be created.}

\item{writeSchema}{ws}

\item{writePrefix}{wp}
}
\value{
a `DuckDB` source.
}
\description{
Create a `DuckDB` source
}
\examples{
library(OmopOnDuckDB)

src <- duckdbSource()

}
