% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{It says hello to whoever you want, by default world.}
\usage{
hello(nam = NULL)
}
\arguments{
\item{nam}{name to say hello}
}
\description{
It says hello to whoever you want, by default world.
}
\examples{
#'
\dontrun{
library(OmopOnDuckDB)
hello()
}
}
