% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopOnDuckDB-package.R
\docType{package}
\name{OmopOnDuckDB-package}
\alias{OmopOnDuckDB}
\alias{OmopOnDuckDB-package}
\title{OmopOnDuckDB: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://OHDSI.github.io/OmopOnDuckDB/}
  \item Report bugs at \url{https://github.com/OHDSI/OmopOnDuckDB/issues}
}

}
\keyword{internal}
