% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{conDoc}
\alias{conDoc}
\title{Helper for consistent documentation of `con`}
\arguments{
\item{con}{A DuckDB connection or path to a duckdb file.}
}
\description{
Helper for consistent documentation of `con`
}
\keyword{internal}
