% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdbSource.R
\name{duckdbSource}
\alias{duckdbSource}
\title{Create a `DuckDB` source}
\usage{
duckdbSource(con = NULL, writeSchema = "main", writePrefix = "")
}
\arguments{
\item{con}{A DuckDB connection or path to a DuckDB database file.}

\item{writeSchema}{Name of the schema where writable tables are created.}

\item{writePrefix}{Prefix applied to writable table names in `writeSchema`.}
}
\value{
a `DuckDB` source.
}
\description{
Create a `DuckDB` source
}
\examples{
library(OmopOnDuckDB)

src <- duckdbSource()

}
